[Pushsafer] Refactor test layout to provide room for expansion v1 vs. v2
